Add new C++ solutions for various problems, update existing implementations, and adjust compiler settings for C++20

+++ Newbies/Chap4/299pndelnum3.cpp
@@ -18,5 +18,5 @@ int main()
         res.pop_back();
         k--;
     }
-    cout << stoi(res);
+    cout << stoll(res);
 }
